+++ client/app/components/Form/index.jsx
@@ -1,24 +1,20 @@
 // @flow
 import React from 'react';
-import { Input } from '../Input';
-import { TYPES as INPUT_TYPES } from '../Input/utils';
-import type { Props as InputProps } from '../Input/utils';
 import { REQUIRES_DEFAULT } from '../Input/InputDefault';
-import { displayQuickCreate } from './QuickCreate';
+import { displayQuickCreate } from './quickCreate';
+import { displayInputs } from '../../widgets/Inputs';
+import { TYPES as INPUT_TYPES } from '../Input/utils';
 import { Utils } from '../../utils';
 import css from './Form.scss';
 
-type KeyProps = { myKey?: any };
-
-type MyInputProps = InputProps & KeyProps;
 
 type Errors = { [string]: boolean } | {};
 
 export type Props = {
   action?: string,
   inputs: any[],
-  noFormTag?: boolean, // Can't have nested forms i.e. quick create
   noFormTagSubmit?: Function,
+  noFormTag?: boolean, // Can't have nested forms i.e. quick create
   noFormTagRef?: any,
 };
 
@@ -84,67 +80,14 @@ export class Form extends React.Component<Props, State> {
     this.onSubmit(e);
   };
 
-  displayInput = (input: MyInputProps) => {
-    const { noFormTag } = this.props;
-    return (
-      <div key={input.id}>
-        <Input
-          id={input.id}
-          key={input.myKey}
-          type={input.type}
-          name={input.name}
-          label={input.label}
-          placeholder={input.placeholder}
-          error={input.error}
-          value={input.value}
-          readOnly={input.readOnly}
-          copyOnClick={input.copyOnClick}
-          disabled={input.disabled}
-          required={input.required}
-          info={input.info}
-          minLength={input.minLength}
-          maxLength={input.maxLength}
-          dark={input.dark}
-          large={input.large}
-          checked={input.checked}
-          uncheckedValue={input.uncheckedValue}
-          options={input.options}
-          checkboxes={input.checkboxes}
-          accordion={input.accordion}
-          onClick={
-            input.type === 'submit' && noFormTag
-              ? this.handleNoFormTagSubmit
-              : undefined
-          }
-          onError={input.type !== 'submit' ? this.handleError : undefined}
-          myRef={(element) => {
-            this.myRefs[input.id] = element;
-          }}
-          formNoValidate={input.type === 'submit'}
-        />
-      </div>
-    );
-  };
 
-  displayInputs = (): any => {
-    const { inputs } = this.state;
-    return inputs.map((input: any) => {
-      if (INPUT_TYPES.includes(input.type)) {
-        return this.displayInput(input);
-      }
-      if (input.type === 'quickCreate') {
-        return displayQuickCreate(input);
-      }
-      return null;
-    });
-  };
 
   render() {
     const { action, noFormTag, noFormTagRef } = this.props;
     if (noFormTag) {
       return (
         <div className={css.form} ref={noFormTagRef}>
-          {this.displayInputs()}
+          {displayInputs()}
         </div>
       );
     }
@@ -157,7 +100,7 @@ export class Form extends React.Component<Props, State> {
         method="post"
         action={action}
       >
-        {this.displayInputs()}
+        {displayInputs()}
       </form>
     );
   }
